libdeflate compiler warnings

+++ 2.0/libdeflate/lib/arm/arm_cpu_features.c
@@ -70,7 +70,7 @@ static void scan_auxv(unsigned long *hwcap, unsigned long *hwcap2)
 				goto out;
 			}
 			filled += ret;
-		} while (filled < 2 * sizeof(long));
+		} while ((unsigned int)filled < 2 * sizeof(long));
 
 		i = 0;
 		do {
@@ -83,7 +83,7 @@ static void scan_auxv(unsigned long *hwcap, unsigned long *hwcap2)
 				*hwcap2 = value;
 			i += 2;
 			filled -= 2 * sizeof(long);
-		} while (filled >= 2 * sizeof(long));
+		} while ((unsigned int)filled >= 2 * sizeof(long));
 
 		memmove(auxbuf, &auxbuf[i], filled);
 	}
